Board rev 2
Different GPIO's used and lower 1.8 VCC voltage. The lower supply
voltage requires writing to the flash byte-wise and also lowering
the MCU frequency to less than 180Mhz. We set it to 144Mhz for now.
Use a different PLL setting for 1.8v core. The previous setting was
mathematically correct but resulted in a clock that was of by >20%
for unknown reasons

The makefile now outputs different targets for different board
revisions although object files for different board revisions
are not differentiated so a clean needs to be done when changing
to build for a different board